+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+### Changed
+- Changed property names for specifying custom claims (see configuration template for required format)
+- UserInfo modifiers are now loaded only at the startup, previously were loaded for each modification separately
+### Fixed
+- Fixed possible null pointer exceptions and wrong behavior for FilterEduPersonEntitlement UserInfo modifier
 
 ## [v1.24.0]
 ### Added