+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Added service name to the consent header
+### Changed
+- Removed reference to CERIT-SC from CESNET footer
 
 ## [v1.24.0]
 ### Added